refactor: 패키지 양방향 의존 제거
