+++ tomcat/src/main/java/nextstep/jwp/handler/LoginHandler.java
@@ -1,7 +1,8 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.handler;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.catalina.servletcontainer.Handler;
 import org.apache.coyote.http11.common.Cookie;
 import org.apache.coyote.http11.common.FileReader;
 import org.apache.coyote.http11.common.Session;
