+++ tomcat/src/main/java/org/apache/coyote/http11/Container.java
@@ -1,9 +1,9 @@
-package org.apache.coyote.http11.handler;
+package org.apache.coyote.http11;
 
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
-public interface Handler {
+public interface Container {
 
     void handle(HttpRequest httpRequest, HttpResponse httpResponse) throws Exception;
 }
