+++ tomcat/src/test/java/nextstep/jwp/handler/LoginHandlerTest.java
@@ -1,9 +1,10 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.handler;
 
 import org.apache.coyote.http11.common.Cookie;
 import org.apache.coyote.http11.common.Session;
 import org.apache.coyote.http11.common.SessionManager;
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.RequestParser;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.junit.jupiter.api.Test;
