+++ tomcat/src/test/java/nextstep/jwp/handler/RegisterHandlerTest.java
@@ -1,6 +1,7 @@
-package org.apache.coyote.http11.handler;
+package nextstep.jwp.handler;
 
 import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.request.RequestParser;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.junit.jupiter.api.Test;
