+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestHeaderTest.java
@@ -1,7 +1,6 @@
 package org.apache.coyote.http11.request;
 
 import org.apache.coyote.http11.common.Cookie;
-import org.apache.coyote.http11.handler.RequestParser;
 import org.junit.jupiter.api.Test;
 
 import java.io.BufferedReader;
